feat: Add `Disclosure` module and deprecate `Hidden`
