+++ docs/BASIC_CONCEPTS.md
@@ -12,11 +12,11 @@ Reakit is established on principles that make it more consistent throughout its 
 
 ## Components
 
-Like any other component library, components are the **highest level API** in Reakit. The [Hidden](/docs/hidden/) component, for example, renders an element that can be hidden or visible.
+Like any other component library, components are the **highest level API** in Reakit. The [DisclosureContent](/docs/disclosure/) component, for example, renders an element that can be hidden or visible.
 
 <!-- eslint-disable -->
 ```jsx static
-<Hidden visible>Hidden</Hidden>
+<DisclosureContent visible>Content</DisclosureContent>
 ```
 
 ## Options
@@ -27,7 +27,7 @@ Components receive two kinds of props: **HTML props** and **option props**. Opti
 ```jsx static
 // `className` is an HTML prop
 // `visible` is an option
-<Hidden className="class" visible />
+<DisclosureContent className="class" visible />
 ```
 
 ## `as` prop
@@ -35,13 +35,13 @@ Components receive two kinds of props: **HTML props** and **option props**. Opti
 Components render only one element. You can change its type using the `as` prop:
 
 ```jsx
-import { Hidden } from "reakit";
+import { DisclosureContent } from "reakit";
 
 function Example() {
   return (
-    <Hidden visible as="button">
-      Hidden Button
-    </Hidden>
+    <DisclosureContent visible as="button">
+      Content
+    </DisclosureContent>
   );
 }
 ```
@@ -53,13 +53,13 @@ Learn more in [Composition](/docs/composition/#as-prop).
 Alternatively, you can change the underlying element by passing children as a function (also known as [render props](https://reactjs.org/docs/render-props.html)):
 
 ```jsx
-import { Hidden, Button } from "reakit";
+import { DisclosureContent, Button } from "reakit";
 
 function Example() {
   return (
-    <Hidden visible>
-      {props => <Button {...props}>Hidden Button</Button>}
-    </Hidden>
+    <DisclosureContent visible>
+      {props => <Button {...props}>Content</Button>}
+    </DisclosureContent>
   );
 }
 ```
@@ -73,15 +73,15 @@ Many Reakit components accept state props, and you can plug your own. As a conve
 The returned options can be passed as props directly to the components, or used separately to access, update and/or [extend the state](/docs/composition/#state-hooks).
 
 ```jsx
-import { useHiddenState, Hidden } from "reakit";
+import { useDisclosureState, DisclosureContent } from "reakit";
 
 function Example() {
   // exposes `visible` state and methods like `show`, `hide` and `toggle`
-  const hidden = useHiddenState({ visible: true });
+  const hidden = useDisclosureState({ visible: true });
   return (
     <>
       <button onClick={hidden.toggle}>Disclosure</button>
-      <Hidden {...hidden}>Hidden</Hidden>
+      <DisclosureContent {...hidden}>Content</DisclosureContent>
     </>
   );
 }
@@ -91,21 +91,26 @@ Learn more in [Managing state](/docs/managing-state/).
 
 ## Props hooks
 
-Finally, as the **lowest level API**, Reakit exposes props hooks. These hooks hold most of the logic behind components and are heavily used within Reakit's source code as a means to compose behaviors without the hassle of polluting the tree with multiple components. For example, [Dialog](/docs/dialog/) uses [Hidden](/docs/hidden/), which in turn uses [Box](/docs/box/).
+Finally, as the **lowest level API**, Reakit exposes props hooks. These hooks hold most of the logic behind components and are heavily used within Reakit's source code as a means to compose behaviors without the hassle of polluting the tree with multiple components. For example, [Dialog](/docs/dialog/) uses [DisclosureContent](/docs/disclosure/), which in turn uses [Box](/docs/box/).
 
 ```jsx
-import { Box, useHiddenState, useHidden, useHiddenDisclosure } from "reakit";
+import {
+  Box,
+  useDisclosureState,
+  useDisclosureContent,
+  useDisclosure
+} from "reakit";
 
 function Example() {
-  const state = useHiddenState({ visible: true });
-  const props = useHidden(state);
-  const disclosureProps = useHiddenDisclosure(state);
+  const state = useDisclosureState({ visible: true });
+  const contentProps = useDisclosureContent(state);
+  const disclosureProps = useDisclosure(state);
   return (
     <>
       <Box as="button" {...disclosureProps}>
         Disclosure
       </Box>
-      <Box {...props}>Hidden</Box>
+      <Box {...contentProps}>Content</Box>
     </>
   );
 }
