+++ docs/MANAGING_STATE.md
@@ -10,13 +10,13 @@ redirect_from:
 
 Reakit components are pretty much stateless, which means that they need props to be passed in so as to modify their state.
 
-[Hidden](/docs/hidden/), for example, is a generic component that can be hidden or visible based on props:
+[DisclosureContent](/docs/disclosure/), for example, is a generic component that can be disclosure or visible based on props:
 
 ```jsx
-import { Hidden } from "reakit";
+import { DisclosureContent } from "reakit";
 
 function Example() {
-  return <Hidden visible>Hidden</Hidden>;
+  return <DisclosureContent visible>Content</DisclosureContent>;
 }
 ```
 
@@ -24,14 +24,14 @@ You can pass your own state to control it:
 
 ```jsx
 import React from "react";
-import { Hidden } from "reakit";
+import { DisclosureContent } from "reakit";
 
 function Example() {
   const [visible, setVisible] = React.useState(true);
   return (
     <>
       <button onClick={() => setVisible(!visible)}>Disclosure</button>
-      <Hidden visible={visible}>Hidden</Hidden>
+      <DisclosureContent visible={visible}>Content</DisclosureContent>
     </>
   );
 }
@@ -44,14 +44,14 @@ As a convenience — and because some states need more complex logic —, Reakit
 The returned [options](/docs/basic-concepts/#options) can be passed as props directly to the components, or used separately to access, update and/or [extend the state](/docs/composition/#state-hooks).
 
 ```jsx
-import { useHiddenState, Hidden } from "reakit";
+import { useDisclosureState, DisclosureContent } from "reakit";
 
 function Example() {
-  const hidden = useHiddenState({ visible: true });
+  const disclosure = useDisclosureState({ visible: true });
   return (
     <>
-      <button onClick={hidden.toggle}>Disclosure</button>
-      <Hidden {...hidden}>Hidden</Hidden>
+      <button onClick={disclosure.toggle}>Disclosure</button>
+      <DisclosureContent {...disclosure}>Content</DisclosureContent>
     </>
   );
 }
@@ -63,7 +63,7 @@ For lazy initialization, you can pass a function that returns the initial state:
 
 <!-- eslint-disable no-undef -->
 ```js
-useHiddenState(() => ({ visible: getExpensiveValue() }));
+useDisclosureState(() => ({ visible: getExpensiveValue() }));
 ```
 
 ## Shared state
@@ -71,27 +71,27 @@ useHiddenState(() => ({ visible: getExpensiveValue() }));
 If you need to share state between multiple components within your app, you can use [Constate](https://github.com/diegohaz/constate):
 
 ```jsx
-import { useHiddenState, Hidden, HiddenDisclosure } from "reakit";
+import { useDisclosureState, Disclosure, DisclosureContent } from "reakit";
 import constate from "constate";
 
-const [HiddenProvider, useHiddenContext] = constate(useHiddenState);
+const [DisclosureProvider, useDisclosureContext] = constate(useDisclosureState);
 
-function Disclosure() {
-  const hidden = useHiddenContext();
-  return <HiddenDisclosure {...hidden}>Disclosure</HiddenDisclosure>;
+function Button() {
+  const disclosure = useDisclosureContext();
+  return <Disclosure {...disclosure}>Disclosure</Disclosure>;
 }
 
-function HiddenElement() {
-  const hidden = useHiddenContext();
-  return <Hidden {...hidden}>Hidden</Hidden>;
+function Content() {
+  const disclosure = useDisclosureContext();
+  return <DisclosureContent {...disclosure}>Content</DisclosureContent>;
 }
 
 function Example() {
   return (
-    <HiddenProvider visible>
-      <Disclosure />
-      <HiddenElement />
-    </HiddenProvider>
+    <DisclosureProvider visible>
+      <Button />
+      <Content />
+    </DisclosureProvider>
   );
 }
 ```
