+++ packages/reakit-playground/src/__deps/reakit.ts
@@ -81,6 +81,10 @@ export default {
   "reakit/Form/FormGroup": require("reakit/Form/FormGroup"),
   "reakit/Form/FormCheckbox": require("reakit/Form/FormCheckbox"),
   "reakit/Form/Form": require("reakit/Form/Form"),
+  "reakit/Disclosure": require("reakit/Disclosure"),
+  "reakit/Disclosure/DisclosureState": require("reakit/Disclosure/DisclosureState"),
+  "reakit/Disclosure/DisclosureContent": require("reakit/Disclosure/DisclosureContent"),
+  "reakit/Disclosure/Disclosure": require("reakit/Disclosure/Disclosure"),
   "reakit/Dialog": require("reakit/Dialog"),
   "reakit/Dialog/DialogState": require("reakit/Dialog/DialogState"),
   "reakit/Dialog/DialogDisclosure": require("reakit/Dialog/DialogDisclosure"),
