+++ packages/reakit/.gitignore
@@ -20,6 +20,7 @@
 /Hidden
 /Group
 /Form
+/Disclosure
 /Dialog
 /Checkbox
 /Button
