+++ packages/reakit/src/Button/README.md
@@ -76,7 +76,7 @@ Learn more in [Accessibility](/docs/accessibility/).
 
 ## Composition
 
-- `Button` uses [Tabbable](/docs/tabbable/), and is used by [FormPushButton](/docs/form/), [FormRemoveButton](/docs/form/), [HiddenDisclosure](/docs/hidden/) and all their derivatives.
+- `Button` uses [Tabbable](/docs/tabbable/), and is used by [FormPushButton](/docs/form/), [FormRemoveButton](/docs/form/), [Disclosure](/docs/disclosure/) and all their derivatives.
 
 Learn more in [Composition](/docs/composition/#props-hooks).
