+++ packages/reakit/src/Dialog/DialogDisclosure.ts
@@ -1,15 +1,15 @@
 import { createComponent } from "reakit-system/createComponent";
 import { createHook } from "reakit-system/createHook";
 import {
-  HiddenDisclosureOptions,
-  HiddenDisclosureHTMLProps,
-  useHiddenDisclosure
-} from "../Hidden/HiddenDisclosure";
+  DisclosureOptions,
+  DisclosureHTMLProps,
+  useDisclosure
+} from "../Disclosure/Disclosure";
 import { useDialogState } from "./DialogState";
 
-export type DialogDisclosureOptions = HiddenDisclosureOptions;
+export type DialogDisclosureOptions = DisclosureOptions;
 
-export type DialogDisclosureHTMLProps = HiddenDisclosureHTMLProps;
+export type DialogDisclosureHTMLProps = DisclosureHTMLProps;
 
 export type DialogDisclosureProps = DialogDisclosureOptions &
   DialogDisclosureHTMLProps;
@@ -19,7 +19,7 @@ export const useDialogDisclosure = createHook<
   DialogDisclosureHTMLProps
 >({
   name: "DialogDisclosure",
-  compose: useHiddenDisclosure,
+  compose: useDisclosure,
   useState: useDialogState,
 
   useProps(_, htmlProps) {
