+++ packages/reakit/src/Disclosure/index.ts
@@ -1,0 +1,3 @@
+export * from "./Disclosure";
+export * from "./DisclosureContent";
+export * from "./DisclosureState";
