+++ packages/reakit/src/Hidden/Hidden.ts
@@ -1,78 +1,32 @@
-import * as React from "react";
 import { createComponent } from "reakit-system/createComponent";
 import { createHook } from "reakit-system/createHook";
-import { cx } from "reakit-utils/cx";
-import { useAllCallbacks } from "reakit-utils/useAllCallbacks";
+import { warning } from "reakit-utils/warning";
 import {
-  unstable_IdGroupOptions,
-  unstable_IdGroupHTMLProps,
-  unstable_useIdGroup
-} from "../Id/IdGroup";
-import { useHiddenState, HiddenStateReturn } from "./HiddenState";
-import { useWarningIfMultiple } from "./__utils/useWarningIfMultiple";
-import { useSetIsMounted } from "./__utils/useSetIsMounted";
+  DisclosureContentOptions,
+  DisclosureContentHTMLProps,
+  useDisclosureContent
+} from "../Disclosure/DisclosureContent";
+import { useHiddenState } from "./HiddenState";
 
-export type HiddenOptions = unstable_IdGroupOptions &
-  Pick<
-    Partial<HiddenStateReturn>,
-    | "visible"
-    | "unstable_animating"
-    | "unstable_animated"
-    | "unstable_stopAnimation"
-    | "unstable_setIsMounted"
-  >;
+export type HiddenOptions = DisclosureContentOptions;
 
-export type HiddenHTMLProps = unstable_IdGroupHTMLProps;
+export type HiddenHTMLProps = DisclosureContentHTMLProps;
 
 export type HiddenProps = HiddenOptions & HiddenHTMLProps;
 
 export const useHidden = createHook<HiddenOptions, HiddenHTMLProps>({
   name: "Hidden",
-  compose: unstable_useIdGroup,
+  compose: useDisclosureContent,
   useState: useHiddenState,
 
-  useProps(
-    options,
-    {
-      onAnimationEnd: htmlOnAnimationEnd,
-      onTransitionEnd: htmlOnTransitionEnd,
-      className: htmlClassName,
-      style: htmlStyle,
-      ...htmlProps
-    }
-  ) {
-    const [hiddenClass, setHiddenClass] = React.useState<string | null>(null);
-
-    useWarningIfMultiple(options);
-    useSetIsMounted(options);
-
-    React.useEffect(() => {
-      setHiddenClass(!options.visible ? "hidden" : null);
-    }, [options.visible]);
-
-    const onTransitionEnd = React.useCallback(() => {
-      if (options.unstable_animated && options.unstable_stopAnimation) {
-        options.unstable_stopAnimation();
-      }
-    }, [options.unstable_animated, options.unstable_stopAnimation]);
-
-    const animating = options.unstable_animated && options.unstable_animating;
-    const hidden = !options.visible && !animating;
-
-    return {
-      role: "region",
-      id: options.baseId,
-      className: cx(hiddenClass, htmlClassName),
-      onAnimationEnd: useAllCallbacks(onTransitionEnd, htmlOnAnimationEnd),
-      onTransitionEnd: useAllCallbacks(onTransitionEnd, htmlOnTransitionEnd),
-      hidden,
-      ...(hidden
-        ? { style: { display: "none", ...htmlStyle } }
-        : htmlStyle
-        ? { style: htmlStyle }
-        : {}),
-      ...htmlProps
-    };
+  useProps(_, htmlProps) {
+    warning(
+      true,
+      "[reakit/Hidden]",
+      "`Hidden` has been renamed to `DisclosureContent`. Using `<Hidden />` will no longer work in future versions.",
+      "See https://reakit.io/docs/disclosure"
+    );
+    return htmlProps;
   }
 });
