+++ packages/reakit/src/Hidden/__tests__/HiddenDisclosure-test.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { render, click } from "reakit-test-utils";
+import { render } from "reakit-test-utils";
 import { HiddenDisclosure } from "../HiddenDisclosure";
 
 const props: Parameters<typeof HiddenDisclosure>[0] = {
@@ -8,62 +8,6 @@ const props: Parameters<typeof HiddenDisclosure>[0] = {
 };
 
 test("render", () => {
-  const { getByText } = render(
-    <HiddenDisclosure {...props}>disclosure</HiddenDisclosure>
-  );
-  expect(getByText("disclosure")).toMatchInlineSnapshot(`
-    <button
-      aria-controls="test"
-      aria-expanded="false"
-      type="button"
-    >
-      disclosure
-    </button>
-  `);
-});
-
-test("render visible", () => {
-  const { getByText } = render(
-    <HiddenDisclosure {...props} visible>
-      disclosure
-    </HiddenDisclosure>
-  );
-  expect(getByText("disclosure")).toMatchInlineSnapshot(`
-    <button
-      aria-controls="test"
-      aria-expanded="true"
-      type="button"
-    >
-      disclosure
-    </button>
-  `);
-});
-
-test("render with aria-controls", () => {
-  const { getByText } = render(
-    <HiddenDisclosure {...props} aria-controls="a">
-      disclosure
-    </HiddenDisclosure>
-  );
-  expect(getByText("disclosure")).toMatchInlineSnapshot(`
-    <button
-      aria-controls="a test"
-      aria-expanded="false"
-      type="button"
-    >
-      disclosure
-    </button>
-  `);
-});
-
-test("toggle", () => {
-  const fn = jest.fn();
-  const { getByText } = render(
-    <HiddenDisclosure {...props} toggle={fn}>
-      disclosure
-    </HiddenDisclosure>
-  );
-  expect(fn).toHaveBeenCalledTimes(0);
-  click(getByText("disclosure"));
-  expect(fn).toHaveBeenCalledTimes(1);
+  render(<HiddenDisclosure {...props}>disclosure</HiddenDisclosure>);
+  expect(console).toHaveWarned();
 });
